feat: return event objects on create/update calls
This is instead of returning true on success or false on failure. In
the case of failures (due to an etag precondition failing), nil is
returned.

The event objects that get returned on success only have the URL and
the etag set - not the calendar data (as this isn't returned in the
response). This change is so the etag can be tracked and kept without
needing to re-request the event object.